--- apps/dolly-frontend/src/main/js/src/components/SearchContainer/index.ts
@@ -1,3 +1,0 @@
-import SearchContainer from './SearchContainer'
-
-export default SearchContainer
